# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FixMyCity Civic Issue Reporter API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type IssueStatus = (typeof IssueStatus)[keyof typeof IssueStatus];

export const IssueStatus = {
  reported: "reported",
  "in-progress": "in-progress",
  resolved: "resolved",
} as const;

export interface Issue {
  id: number;
  title: string;
  category: string;
  description: string;
  digiPin: string;
  location: string;
  reporterName: string;
  reporterEmail: string;
  reporterContact: string;
  status: IssueStatus;
  reportCount: number;
  adminNotes?: string | null;
  submittedAt: string;
  resolvedAt?: string | null;
}

export interface CreateIssueRequest {
  title: string;
  category: string;
  description: string;
  digiPin: string;
  location: string;
  reporterName: string;
  reporterEmail: string;
  reporterContact: string;
}

export type UpdateStatusRequestStatus =
  (typeof UpdateStatusRequestStatus)[keyof typeof UpdateStatusRequestStatus];

export const UpdateStatusRequestStatus = {
  reported: "reported",
  "in-progress": "in-progress",
  resolved: "resolved",
} as const;

export interface UpdateStatusRequest {
  status: UpdateStatusRequestStatus;
  adminNotes?: string;
}

export interface IssueListResponse {
  issues: Issue[];
  total: number;
}

export interface LeaderboardEntry {
  rank: number;
  reporterName: string;
  reporterEmail: string;
  reportCount: number;
  resolvedCount: number;
  points: number;
}

export interface LeaderboardResponse {
  entries: LeaderboardEntry[];
}

export type StatsResponseCategoryBreakdownItem = {
  category: string;
  count: number;
};

export interface StatsResponse {
  totalReports: number;
  resolvedCount: number;
  inProgressCount: number;
  reportedCount: number;
  categoryBreakdown: StatsResponseCategoryBreakdownItem[];
}

export interface AdminLoginRequest {
  username: string;
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  token: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GetIssuesParams = {
  status?: GetIssuesStatus;
  category?: string;
  limit?: number;
  offset?: number;
};

export type GetIssuesStatus =
  (typeof GetIssuesStatus)[keyof typeof GetIssuesStatus];

export const GetIssuesStatus = {
  reported: "reported",
  "in-progress": "in-progress",
  resolved: "resolved",
} as const;

export type AdminGetAllIssuesParams = {
  status?: string;
  category?: string;
};