# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FixMyCity Civic Issue Reporter API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get all civic issues (community feed)
 */
export const GetIssuesQueryParams = zod.object({
  status: zod.enum(["reported", "in-progress", "resolved"]).optional(),
  category: zod.coerce.string().optional(),
  limit: zod.coerce.number().optional(),
  offset: zod.coerce.number().optional(),
});

export const GetIssuesResponse = zod.object({
  issues: zod.array(
    zod.object({
      id: zod.number(),
      title: zod.string(),
      category: zod.string(),
      description: zod.string(),
      digiPin: zod.string(),
      location: zod.string(),
      reporterName: zod.string(),
      reporterEmail: zod.string(),
      reporterContact: zod.string(),
      status: zod.enum(["reported", "in-progress", "resolved"]),
      reportCount: zod.number(),
      adminNotes: zod.string().nullish(),
      submittedAt: zod.string(),
      resolvedAt: zod.string().nullish(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Report a new civic issue
 */
export const CreateIssueBody = zod.object({
  title: zod.string(),
  category: zod.string(),
  description: zod.string(),
  digiPin: zod.string(),
  location: zod.string(),
  reporterName: zod.string(),
  reporterEmail: zod.string(),
  reporterContact: zod.string(),
});

/**
 * @summary Get a single issue by ID
 */
export const GetIssueByIdParams = zod.object({
  id: zod.coerce.number(),
});

export const GetIssueByIdResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  category: zod.string(),
  description: zod.string(),
  digiPin: zod.string(),
  location: zod.string(),
  reporterName: zod.string(),
  reporterEmail: zod.string(),
  reporterContact: zod.string(),
  status: zod.enum(["reported", "in-progress", "resolved"]),
  reportCount: zod.number(),
  adminNotes: zod.string().nullish(),
  submittedAt: zod.string(),
  resolvedAt: zod.string().nullish(),
});

/**
 * @summary Upvote (increment report count) for an existing issue
 */
export const UpvoteIssueParams = zod.object({
  id: zod.coerce.number(),
});

export const UpvoteIssueResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  category: zod.string(),
  description: zod.string(),
  digiPin: zod.string(),
  location: zod.string(),
  reporterName: zod.string(),
  reporterEmail: zod.string(),
  reporterContact: zod.string(),
  status: zod.enum(["reported", "in-progress", "resolved"]),
  reportCount: zod.number(),
  adminNotes: zod.string().nullish(),
  submittedAt: zod.string(),
  resolvedAt: zod.string().nullish(),
});

/**
 * @summary Admin - Update issue status
 */
export const UpdateIssueStatusParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateIssueStatusBody = zod.object({
  status: zod.enum(["reported", "in-progress", "resolved"]),
  adminNotes: zod.string().optional(),
});

export const UpdateIssueStatusResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  category: zod.string(),
  description: zod.string(),
  digiPin: zod.string(),
  location: zod.string(),
  reporterName: zod.string(),
  reporterEmail: zod.string(),
  reporterContact: zod.string(),
  status: zod.enum(["reported", "in-progress", "resolved"]),
  reportCount: zod.number(),
  adminNotes: zod.string().nullish(),
  submittedAt: zod.string(),
  resolvedAt: zod.string().nullish(),
});

/**
 * @summary Get user leaderboard ranked by report count
 */
export const GetLeaderboardResponse = zod.object({
  entries: zod.array(
    zod.object({
      rank: zod.number(),
      reporterName: zod.string(),
      reporterEmail: zod.string(),
      reportCount: zod.number(),
      resolvedCount: zod.number(),
      points: zod.number(),
    }),
  ),
});

/**
 * @summary Get aggregate stats (total reports, resolved count, categories)
 */
export const GetStatsResponse = zod.object({
  totalReports: zod.number(),
  resolvedCount: zod.number(),
  inProgressCount: zod.number(),
  reportedCount: zod.number(),
  categoryBreakdown: zod.array(
    zod.object({
      category: zod.string(),
      count: zod.number(),
    }),
  ),
});

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  success: zod.boolean(),
  token: zod.string(),
});

/**
 * @summary Admin - Get all issues with full details
 */
export const AdminGetAllIssuesQueryParams = zod.object({
  status: zod.coerce.string().optional(),
  category: zod.coerce.string().optional(),
});

export const AdminGetAllIssuesResponse = zod.object({
  issues: zod.array(
    zod.object({
      id: zod.number(),
      title: zod.string(),
      category: zod.string(),
      description: zod.string(),
      digiPin: zod.string(),
      location: zod.string(),
      reporterName: zod.string(),
      reporterEmail: zod.string(),
      reporterContact: zod.string(),
      status: zod.enum(["reported", "in-progress", "resolved"]),
      reportCount: zod.number(),
      adminNotes: zod.string().nullish(),
      submittedAt: zod.string(),
      resolvedAt: zod.string().nullish(),
    }),
  ),
  total: zod.number(),
});