# lib/api-zod/src/generated/types/createIssueRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FixMyCity Civic Issue Reporter API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateIssueRequest {
  title: string;
  category: string;
  description: string;
  digiPin: string;
  location: string;
  reporterName: string;
  reporterEmail: string;
  reporterContact: string;
}