# lib/api-zod/src/generated/types/issue.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FixMyCity Civic Issue Reporter API
 * OpenAPI spec version: 0.1.0
 */
import type { IssueStatus } from "./issueStatus";

export interface Issue {
  id: number;
  title: string;
  category: string;
  description: string;
  digiPin: string;
  location: string;
  reporterName: string;
  reporterEmail: string;
  reporterContact: string;
  status: IssueStatus;
  reportCount: number;
  adminNotes?: string | null;
  submittedAt: string;
  resolvedAt?: string | null;
}