# lib/api-zod/src/generated/types/issueListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FixMyCity Civic Issue Reporter API
 * OpenAPI spec version: 0.1.0
 */
import type { Issue } from "./issue";

export interface IssueListResponse {
  issues: Issue[];
  total: number;
}