# lib/api-zod/src/generated/types/leaderboardResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FixMyCity Civic Issue Reporter API
 * OpenAPI spec version: 0.1.0
 */
import type { LeaderboardEntry } from "./leaderboardEntry";

export interface LeaderboardResponse {
  entries: LeaderboardEntry[];
}