# lib/api-zod/src/generated/types/statsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FixMyCity Civic Issue Reporter API
 * OpenAPI spec version: 0.1.0
 */
import type { StatsResponseCategoryBreakdownItem } from "./statsResponseCategoryBreakdownItem";

export interface StatsResponse {
  totalReports: number;
  resolvedCount: number;
  inProgressCount: number;
  reportedCount: number;
  categoryBreakdown: StatsResponseCategoryBreakdownItem[];
}